+++ src/control_plane/network/mod.rs
@@ -1,0 +1,16 @@
+mod global_network_reconcile_action;
+mod global_network_reconcile_error;
+mod global_network_reconcile_options;
+mod global_network_reconcile_result;
+mod global_network_request;
+mod reconcile_global_network;
+
+pub(crate) use global_network_reconcile_action::GlobalNetworkReconcileAction;
+pub(crate) use global_network_reconcile_error::GlobalNetworkReconcileError;
+pub(crate) use global_network_reconcile_options::GlobalNetworkReconcileOptions;
+pub(crate) use global_network_reconcile_result::GlobalNetworkReconcileResult;
+pub(crate) use global_network_request::global_network_request;
+pub(crate) use reconcile_global_network::reconcile_global_network;
+
+#[cfg(test)]
+mod tests;